More progress - treatments are tested
