+++ .travis.yml
@@ -18,4 +18,4 @@ install:
 - pip install -r requirements_test.txt
 
 script:
-- py.test --cov taar_lite tests
+- py.test --cov taar_lite
